Catch illegal `this` invocations
If you called any of the native methods with an unexpected `this`
argument it would just fail an assertion at best, or continue with some
other object at worst. This adds an instance check that you can't fake
in plain Javascript to avoid those issues. Now Unwrap will throw to the
calling code if there's a problem.