Fix failure when multiple fibers finish together
If multiple fibers finished one after in an unexpected way there would
be a memory leak because `delete_me` would get overwritten. There was an
assertion in place to prevent this scenario, and a case was discovered
which triggers this condition. It was actually quite hard to hit because
the coroutine pool masks this issue very well. The issue was fixed by
deleting dead coroutines as soon as possible, instead of at the next
call to run()

Closes gh-16